chore: improve task graph and test coverage
