+++ packages/notifications/package.json
@@ -32,7 +32,8 @@
     "build": "tsc",
     "check": "tsc --pretty --noEmit",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "dev": "tsc --watch"
+    "dev": "tsc --watch",
+    "test": "vitest run"
   },
   "dependencies": {
     "@heroicons/react": "2.0.13",
@@ -47,7 +48,8 @@
     "@types/react": "catalog:react",
     "@types/react-dom": "catalog:react",
     "sushi": "catalog:web3",
-    "typescript": "catalog:"
+    "typescript": "catalog:",
+    "vitest": "2.1.9"
   },
   "peerDependencies": {
     "react": "catalog:react",
