--- packages/sushiswap/README.md
@@ -1,2 +1,0 @@
-# SushiSwap
-
